feat(v2.2.0): Milestone v2.2 release — Aufräumen, WebDAV-Export & Wochentag-Muster
Phases 43-48 milestone deliverables (16 Requirements):
- SDF-03/04/05: Special-Days-Feintuning (Kalenderjahr-Loader-Fix,
  Duplikat-Hinweis Replace-Copy, Feiertag<->Kurzer-Tag atomarer Roundtrip)
- BUG-01/02/03: Frontend-Korrektheit (save_slot_edit snapshot-vor-await,
  invitation-parse visible, BackdropPress invariant)
- HYG-03: shifty-dioxus Clippy 177 -> 0, FE-Gate scharfgestellt
- HYG-04/05, IMP-05: Backend-Hygiene & i18n (Edit-structure-Keys, REST
  content-type surface drift-guard, i18n_impersonation-Test grün)
- RPT-01/02/03: Wochentag-Anwesenheits-Muster (pure fn
  weekday_attendance_distribution, neuer Endpoint-Shape, FE HR-Stats-Row-Swap,
  Snapshot bleibt 12)
- EXP-01/02/03: Nextcloud-PDF-Export via WebDAV (printpdf-Renderer,
  reqwest_dav-Client mit Retry-Loop 2s/4s/8s, tokio-cron-scheduler,
  pdf_export_config-Migration + Basic-Service + REST admin-gated GET/PUT/
  trigger, PdfExportScheduler + FE Settings-Card 4)

Post-Ship-Bugfixes (vor Release-Tag):
- SDF-03 DB-Filter: get_by_year lädt year + year-1, filtert per Kalenderdatum-
  Jahr, sortiert nach Datum (Fix: 01.01.YYYY war unter ISO-Wochenjahr YYYY-1
  gespeichert und im YYYY-Filter unsichtbar)
- RPT-02 Sichtbarkeit: is_dynamic-Gate im Backend entfernt (HR-Gate bleibt) +
  neuer kurzer WeekdayAttendanceLabel-i18n-Key als TupleRow-Label

Neue Dependencies: printpdf, reqwest_dav, tokio-cron-scheduler, base64,
thiserror (prod) + wiremock (test). Eine neue Migration
(pdf_export_config Single-Row). Snapshot bleibt 12 (kein Bump).

Alle Gates grün: Backend cargo test --workspace, cargo clippy --workspace
-- -D warnings, cargo build --workspace; Frontend cargo build --target
wasm32-unknown-unknown, cargo test -p shifty-dioxus (787 passed), cargo
clippy -p shifty-dioxus -- -D warnings.

Sowie:
- Set version to 2.2.0 (Cargo.toml bumps, nix files, .sqlx/)
- Milestone-Close: MILESTONES.md v2.2-section, PROJECT/STATE/ROADMAP
  updates, milestones/v2.2-{ROADMAP,REQUIREMENTS,MILESTONE-AUDIT}.md
  archives, REQUIREMENTS.md removed

Consolidated into a single commit due to a jj tree-encoding bug in the
original split chain that produced duplicate tree entries and blocked
push (fsck: duplicateEntries). All post-v2.1.0 work is preserved.

+++ shifty-dioxus/src/state/sales_person_available.rs
@@ -17,7 +17,7 @@ impl From<&SalesPersonUnavailableTO> for SalesPersonUnavailable {
             id: sales_person_available.id,
             sales_person_id: sales_person_available.sales_person_id,
             year: sales_person_available.year,
-            week: sales_person_available.calendar_week as u8,
+            week: sales_person_available.calendar_week,
             day_of_week: sales_person_available.day_of_week.into(),
             version: sales_person_available.version,
         }
